refactor: proxy 하는 객체 분리
