+++ src/main/java/persistence/proxy/ProxyFactory.java
@@ -1,0 +1,5 @@
+package persistence.proxy;
+
+public interface ProxyFactory {
+    Object createProxy(Class<?> clazz);
+}
